+++ src/main/java/net/frozenblock/wilderwild/registry/WilderWildRegistries.java
@@ -18,23 +18,22 @@
 
 package net.frozenblock.wilderwild.registry;
 
+import net.fabricmc.fabric.api.event.registry.DynamicRegistries;
 import net.fabricmc.fabric.api.event.registry.FabricRegistryBuilder;
 import net.fabricmc.fabric.api.event.registry.RegistryAttribute;
 import net.frozenblock.wilderwild.WWConstants;
 import net.frozenblock.wilderwild.entity.variant.ButterflyVariant;
-import net.frozenblock.wilderwild.entity.variant.FireflyColor;
+import net.frozenblock.wilderwild.entity.variant.firefly.FireflyColor;
 import net.frozenblock.wilderwild.entity.variant.JellyfishVariant;
 import net.minecraft.core.MappedRegistry;
 import net.minecraft.core.Registry;
 import net.minecraft.resources.ResourceKey;
 
 public final class WilderWildRegistries {
-	public static final ResourceKey<Registry<FireflyColor>> FIREFLY_COLOR_REGISTRY = ResourceKey.createRegistryKey(WWConstants.id("firefly_color"));
+	public static final ResourceKey<Registry<FireflyColor>> FIREFLY_COLOR = ResourceKey.createRegistryKey(WWConstants.id("firefly_color"));
 	public static final ResourceKey<Registry<ButterflyVariant>> BUTTERFLY_VARIANT_REGISTRY = ResourceKey.createRegistryKey(WWConstants.id("butterfly_variant"));
 	public static final ResourceKey<Registry<JellyfishVariant>> JELLYFISH_VARIANT_REGISTRY = ResourceKey.createRegistryKey(WWConstants.id("jellyfish_variant"));
-	public static final MappedRegistry<FireflyColor> FIREFLY_COLOR = FabricRegistryBuilder.createSimple(FIREFLY_COLOR_REGISTRY)
-		.attribute(RegistryAttribute.SYNCED)
-		.buildAndRegister();
+
 	public static final MappedRegistry<ButterflyVariant> BUTTERFLY_VARIANT = FabricRegistryBuilder.createSimple(BUTTERFLY_VARIANT_REGISTRY)
 		.attribute(RegistryAttribute.SYNCED)
 		.buildAndRegister();
@@ -47,7 +46,8 @@ public final class WilderWildRegistries {
 	}
 
 	public static void initRegistry() {
-		FireflyColor.init();
+		DynamicRegistries.registerSynced(FIREFLY_COLOR, FireflyColor.DIRECT_CODEC);
+
 		ButterflyVariant.init();
 		JellyfishVariant.init();
 	}